test(administration): split extension tooling setup spec below the line limit
setup.spec.ts crossed the 500-line ADR threshold; the CLI-entry and
probe-cache scenario groups move into setup.spec/ per
adr/2026-05-06-split-large-administration-test-files.md.
